+++ packages/webblocks-cms/resources/views/pages/partials/blocks/image.blade.php
@@ -1,6 +1,8 @@
 @php
   $image = $block->media;
-  $imageSource = $image?->url();
+  $imageSource = $image?->transformUrl('content');
+  $imageSmallSource = $image?->transformUrl('content-small');
+  $imageLargeSource = $image?->transformUrl('content-large');
   $caption = trim((string) ($block->title ?? ''));
   $altText = trim((string) ($block->subtitle ?? ''));
   $fallbackAltText = trim((string) ($image?->alt_text ?: $image?->title ?: $caption ?: 'Image'));
@@ -22,7 +24,11 @@
     @endif
     <img
       src="{{ $imageSource }}"
+      srcset="{{ $imageSmallSource }} 640w, {{ $imageSource }} 1280w, {{ $imageLargeSource }} 1920w"
+      sizes="(max-width: 800px) 100vw, 1280px"
       alt="{{ $resolvedAltText }}"
+      loading="lazy"
+      decoding="async"
       @if ($image?->width) width="{{ $image->width }}" @endif
       @if ($image?->height) height="{{ $image->height }}" @endif
     >