+++ packages/webblocks-cms/routes/admin.php
@@ -237,6 +237,7 @@ Route::middleware(['web', 'install.required', UseCmsAuthenticationRedirect::clas
     Route::delete('media/bulk', [MediaController::class, 'bulkDestroy'])->name('media.bulk-destroy');
     Route::get('media/{media}', [MediaController::class, 'show'])->name('media.show');
     Route::get('media/{media}/edit', [MediaController::class, 'edit'])->name('media.edit');
+    Route::post('media/{media}/transforms/regenerate', [MediaController::class, 'regenerateTransforms'])->name('media.transforms.regenerate');
     Route::put('media/{media}', [MediaController::class, 'update'])->name('media.update');
     Route::delete('media/{media}', [MediaController::class, 'destroy'])->name('media.destroy');
     Route::post('navigation/reorder', [NavigationItemController::class, 'reorder'])->name('navigation.reorder');